+++ Sources/AudioKit/Nodes/Node.swift
@@ -98,32 +98,7 @@ public extension Node {
 
 extension Node {
 
-    func disconnectAndDetachIfLast(input: Node) {
-        if let engine = avAudioNode.engine {
-            let points = engine.outputConnectionPoints(for: input.avAudioNode, outputBus: 0)
-            let otherConnections = points.filter { $0.node != self.avAudioNode }
-            if otherConnections.isEmpty {
-                // It is important to go depth first search.
-                // If we first detach the current node,
-                // upstream nodes will lose the connection to the engine.
-                for connection in input.connections {
-                    input.disconnectAndDetachIfLast(input: connection)
-                }
-                engine.detach(input.avAudioNode)
-            } else {
-                avAudioNode.disconnect(input: input.avAudioNode)
-            }
-        }
-    }
 
-    func detach() {
-        if let engine = avAudioNode.engine {
-            engine.detach(avAudioNode)
-        }
-        for connection in connections {
-            connection.detach()
-        }
-    }
 
     func disconnectAV() {
         if let engine = avAudioNode.engine {
