+++ Tests/AudioKitTests/Node Tests/NodeTests.swift
@@ -459,20 +459,24 @@ class NodeTests: XCTestCase {
     #endif
 
     func testAllNodesInChainDeallocatedOnRemove() {
-        let engine = AudioEngine()
-        var chain: Node? = createChain()
-        weak var weakPitch = chain?.avAudioNode
-        weak var weakDelay = chain?.connections.first?.avAudioNode
-        weak var weakPlayer = chain?.connections.first?.connections.first?.avAudioNode
-        let mixer = Mixer(chain!, createChain())
-        engine.output = mixer
+        for strategy in [DisconnectStrategy.recursive, .deatach] {
+            let engine = AudioEngine()
+            var chain: Node? = createChain()
+            weak var weakPitch = chain?.avAudioNode
+            weak var weakDelay = chain?.connections.first?.avAudioNode
+            weak var weakPlayer = chain?.connections.first?.connections.first?.avAudioNode
+            let mixer = Mixer(chain!, createChain())
+            engine.output = mixer
 
-        mixer.removeInput(chain!)
-        chain = nil
+            mixer.removeInput(chain!, strategy: strategy)
+            chain = nil
 
-        XCTAssertNil(weakPitch)
-        XCTAssertNil(weakDelay)
-        XCTAssertNil(weakPlayer)
+            XCTAssertNil(weakPitch)
+            XCTAssertNil(weakDelay)
+            XCTAssertNil(weakPlayer)
+
+            XCTAssertFalse(engine.avEngine.description.contains("other nodes"))
+        }
     }
 
     @available(iOS 13.0, *)
@@ -521,23 +525,31 @@ class NodeTests: XCTestCase {
 
     @available(iOS 13.0, *)
     func testInnerNodesThatHaveMultipleInnerConnectionsDeallocated() {
-        let engine = AudioEngine()
-        var chain: Node? = createChain()
-        weak var weakPitch = chain?.avAudioNode
-        weak var weakDelay = chain?.connections.first?.avAudioNode
-        weak var weakPlayer = chain?.connections.first?.connections.first?.avAudioNode
-        var mixer: Mixer? = Mixer(chain!, Mixer(chain!))
-        var outer: Mixer? = Mixer(mixer!)
-        engine.output = outer
+        for strategy in [DisconnectStrategy.recursive, .deatach] {
+            let engine = AudioEngine()
+            var chain: Node? = createChain()
+            weak var weakPitch = chain?.avAudioNode
+            weak var weakDelay = chain?.connections.first?.avAudioNode
+            weak var weakPlayer = chain?.connections.first?.connections.first?.avAudioNode
+            var mixer: Mixer? = Mixer(chain!, Mixer(chain!))
+            var outer: Mixer? = Mixer(mixer!)
+            engine.output = outer
 
-        outer!.removeInput(mixer!)
-        outer = nil
-        mixer = nil
-        chain = nil
+            outer!.removeInput(mixer!, strategy: strategy)
+            outer = nil
+            mixer = nil
+            chain = nil
 
-        XCTAssertNil(weakPitch)
-        XCTAssertNil(weakDelay)
-        XCTAssertNil(weakPlayer)
+            XCTAssertNil(weakPitch)
+            XCTAssertNil(weakDelay)
+            XCTAssertNil(weakPlayer)
+
+            // http://openradar.appspot.com/radar?id=5616162842869760
+            // This condition should be passing, but unfortunately,
+            // under certain conditions, it is not due to a bug.
+
+            // XCTAssertFalse(engine.avEngine.description.contains("other nodes"))
+        }
     }
 }
